+++ src/Queue.php
@@ -16,10 +16,12 @@ use MongoDB\BSON\UTCDateTime;
 use MongoDB\Database;
 use Psr\Log\LoggerInterface;
 use TaskScheduler\Exception\SpawnForkException;
+use League\Event\Emitter;
 
 class Queue
 {
     use InjectTrait;
+    use EventsTrait;
 
     /**
      * Database.
@@ -73,13 +75,14 @@ class Queue
     /**
      * Init queue.
      */
-    public function __construct(Scheduler $scheduler, Database $db, WorkerFactoryInterface $factory, LoggerInterface $logger)
+    public function __construct(Scheduler $scheduler, Database $db, WorkerFactoryInterface $factory, LoggerInterface $logger, Emitter $emitter)
     {
         $this->db = $db;
         $this->logger = $logger;
         $this->factory = $factory;
         $this->jobs = new MessageQueue($db, $scheduler->getJobQueue(), $scheduler->getJobQueueSize(), $logger);
         $this->events = new MessageQueue($db, $scheduler->getEventQueue(), $scheduler->getEventQueueSize(), $logger);
+        $this->emitter = $emitter;
     }
 
     /**
@@ -169,7 +172,7 @@ class Queue
         $cursor_events = $this->events->getCursor([
             'timestamp' => ['$gte' => new UTCDateTime()],
             'job' => ['$exists' => true],
-            'status' => ['$gt' => JobInterface::STATUS_POSTPONED],
+            //'status' => ['$gt' => JobInterface::STATUS_POSTPONED],
         ]);
 
         $this->catchSignal();
@@ -240,11 +243,15 @@ class Queue
      */
     protected function handleEvent(array $event): self
     {
-        $this->logger->debug('received event ['.$event['status'].'] for job ['.$event['job'].'], write into systemv queue', [
-            'category' => get_class($this),
-        ]);
+        $this->emit($this->scheduler->getJob($event['job']));
 
-        msg_send($this->queue, WorkerManager::TYPE_EVENT, $event);
+        if($event['status'] > self::STATUS_POSTPONED) {
+            $this->logger->debug('received event ['.$event['status'].'] for job ['.$event['job'].'], write into systemv queue', [
+                'category' => get_class($this),
+            ]);
+
+            msg_send($this->queue, WorkerManager::TYPE_EVENT, $event);
+        }
 
         return $this;
     }